+++ backend/src/test/java/net/es/oscars/cuke/HoldControllerSteps.java
@@ -110,6 +110,9 @@ public class HoldControllerSteps {
         return generateMockConnection(null);
     }
     private Connection generateMockConnection(String projectId) {
+        Set<String> projectIds = new HashSet<>();
+        projectIds.add(projectId);
+
         return Connection.builder()
             .connectionId("ABCD")
             .phase(Phase.HELD)
@@ -121,7 +124,7 @@ public class HoldControllerSteps {
             .description("test description")
             .connection_mtu(10000)
             .last_modified( ((Long) Instant.now().getEpochSecond()).intValue() )
-            .projectId(null)
+            .projectId(projectIds)
             .build();
     }
     private void setupMockConnSvc() throws Exception {